requirements.txt
python-dotenv==1.0.1
pyyaml==6.0.2
openai==1.59.9